Improve name + number handling
It now disregards capitalized words at the start of a sentence, because
they might not be names. So actual instances of name + number (e.g.
"Battlefield 1") at the start of a sentence will be false negatives, but
that's better than a false positive on a sentence starting with
something like "In 22 weeks".

It can now also detect most Swedish names, such as "Mästerkatten 2".